Cleaning up and restructuring
