+++ api/v2/drivercoverage/surface/presentmode.php
@@ -20,7 +20,7 @@
 	*/
 
 	include './../../../../database/database.class.php';
-	include './../../../../functions.php';
+	include './../../../../includes/functions.php';
 	include './../apiendpoint.php';
 
 	header('Content-Type: application/json');
