+++ compare.php
@@ -22,7 +22,7 @@
 
 require './database/database.class.php';
 require 'page_generator.php';
-require './functions.php';
+require './includes/functions.php';
 require './includes/constants.php';
 require './reportcompare/reportcompare.class.php';
