+++ listmemory.php
@@ -20,7 +20,7 @@
 	
 	include 'page_generator.php';
 	include './database/database.class.php';
-	include './functions.php';
+	include './includes/functions.php';
 
 	$platform = "windows";
 	if (isset($_GET['platform'])) {
